Chrry

+++ apps/api/lib/cleanupTest.ts
@@ -1,21 +1,22 @@
 import {
+  and,
   db,
   deleteCreditUsage,
   deleteInstruction,
   deleteMessage,
   deletePlaceHolder,
-  deleteStore,
   deleteSubscription,
   deleteThread,
   getGuest as getGuestDb,
   getInstructions,
   getMessages,
   getPlaceHolders,
-  getStores,
   getSubscriptions,
   getThreads,
   getUser,
   type guest,
+  isE2E,
+  isProd,
   TEST_GUEST_FINGERPRINTS,
   TEST_MEMBER_EMAILS,
   TEST_MEMBER_FINGERPRINTS,
@@ -26,11 +27,14 @@ import {
   VEX_LIVE_FINGERPRINTS,
 } from "@repo/db"
 import {
+  apps,
   feedbackTransactions,
   GUEST_CREDITS_PER_MONTH,
   MEMBER_CREDITS_PER_MONTH,
   pearFeedback,
+  stores,
 } from "@repo/db/src/schema"
+import { eq, not } from "drizzle-orm"
 
 const allowedFingerprints = TEST_GUEST_FINGERPRINTS.concat(
   TEST_MEMBER_FINGERPRINTS,
@@ -38,6 +42,13 @@ const allowedFingerprints = TEST_GUEST_FINGERPRINTS.concat(
 )
 
 export default async function cleanupTest() {
+  if (isProd) {
+    return { error: "Oops, this is PROD", status: 401 as const }
+  }
+
+  if (!isE2E) {
+    return { error: "Unauthorized", status: 401 as const }
+  }
   // await db.delete(tribePosts)
   await db.delete(pearFeedback)
   await db.delete(feedbackTransactions)
@@ -65,7 +76,7 @@ export default async function cleanupTest() {
       guest.fingerprint === fingerprint &&
       allowedFingerprints.includes(fingerprint)
     ) {
-      await cleanup({
+      return await cleanup({
         guest,
       })
     }
@@ -143,6 +154,12 @@ async function cleanup({ user, guest }: { user?: user; guest?: guest }) {
     }),
   )
 
+  const admin = await getUser({ email: process.env.VEX_TEST_EMAIL })
+
+  if (!admin) {
+    return { success: false, error: "Admin not found", status: 401 as const }
+  }
+
   const placeholders = await getPlaceHolders({
     userId: user?.id,
     guestId: guest?.id,
@@ -161,18 +178,13 @@ async function cleanup({ user, guest }: { user?: user; guest?: guest }) {
     }),
   )
 
-  const stores = await getStores({
-    pageSize: 100000,
-  })
-  await Promise.all(
-    stores.stores.map((store) => {
-      if (store?.user?.email === process.env.VEX_TEST_EMAIL) {
-        return
-      }
+  await db
+    .delete(stores)
+    .where(and(not(eq(stores.userId, admin.id!)), eq(stores.isSystem, false)))
 
-      return deleteStore({ id: store.store.id })
-    }),
-  )
+  await db
+    .delete(apps)
+    .where(and(not(eq(apps.userId, admin.id!)), eq(apps.isSystem, false)))
 
   // 4. Delete subscriptions
   const subscriptions = await getSubscriptions({
@@ -214,6 +226,8 @@ async function cleanup({ user, guest }: { user?: user; guest?: guest }) {
       pearFeedbackCount: 0,
       characterProfilesEnabled: false,
     }))
+
+  return { success: true }
 
   // 5. Clear graph data (FalkorDB)
   // Remove all graph entities and relationships for this user/guest
